logging: Fix timestamp 32bit integer overflow

See merge request wayland/wayland!472

+++ src/connection.c
@@ -1500,7 +1500,7 @@ wl_closure_print(struct wl_closure *closure, struct wl_object *target,
 	struct argument_details arg;
 	const char *signature = closure->message->signature;
 	struct timespec tp;
-	unsigned int time;
+	uint64_t time;
 	uint32_t nval;
 	FILE *f;
 	char *buffer;
@@ -1511,10 +1511,9 @@ wl_closure_print(struct wl_closure *closure, struct wl_object *target,
 		return;
 
 	clock_gettime(CLOCK_REALTIME, &tp);
-	time = (tp.tv_sec * 1000000L) + (tp.tv_nsec / 1000);
+	time = (tp.tv_sec * 1000000LL) + (tp.tv_nsec / 1000);
 
-	fprintf(f, "[%7u.%03u] ", time / 1000, time % 1000);
-
+	fprintf(f, "[%7" PRIu64 ".%03u] ", time / 1000, (unsigned int) (time % 1000));
 	if (queue_name)
 		fprintf(f, "{%s} ", queue_name);
